Move arg_list.h to correct folder
