+++ core/include/scipp/core/element/geometric_operations.h
@@ -5,7 +5,7 @@
 #pragma once
 
 #include "scipp/common/overloaded.h"
-#include "scipp/core/arg_list.h"
+#include "scipp/core/element/arg_list.h"
 #include "scipp/core/transform_common.h"
 
 namespace scipp::core {
